Split ItemViewerContext for feature-flagged viewer refactor
Create legacy.tsx and refactored.tsx (currently identical) from the
original context. The index re-exports both and provides a
useItemViewerContext hook that returns the correct context based on
the itemViewerRefactor toggle. The refactored IIIFViewer uses
ItemViewerContextRefactored as its Provider.
